refactored CourseSectionTableView into its own component to help reduce size of courseSection component
